+++ src/feed/mod.rs
@@ -64,7 +64,7 @@ impl From<RssEntry> for Article {
                 .get(0)
                 .map(|x| x.href.clone())
                 .unwrap_or(entry.id),
-            date: entry.updated.map(DateTime::<Local>::from),
+            date: entry.published.map(DateTime::<Local>::from),
         }
     }
 }