+++ CHANGELOG.md
@@ -15,10 +15,9 @@ Released on ??
   - This should be preferred since contains much more information than just the summary. Unfortunately sometimes the content is empty, so summary should be used as fallback
 - Changed stdlib `Textarea` to `tui-realm-textarea` component to improve readability of summary
 - Fixed the article date shown: was `updated` but should be `published`
-- Migrated UI to tui-realm 1.7.0
+- Migrated UI to tui-realm 1.7.1
 - Updated dependencies
   - argh `0.1.8`
-  - crossterm `0.24`
   - feed-rs `1.1.0`
   - lazy-regex `2.3.0`
   - lazy_static `1.4.0`